External APIs
First and foremost our workflow is currently setup to run each time there is a push event to this repository. As you can imagine, this generates a lot of noise when learning.

Let's comment out our current workflow to prevent things from running but preserve the things you've worked on up to this point.