Create action.yml
Our action does not require much metadata for it to run correctly. We will not be accepting any inputs, we will however be setting a single output this time.

We will not use the fact in in this portion of the course. There will be a later step that will rely on this actions output.